requirements.txt
# Core LLM and Agent Framework
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0

# OpenAI SDK
openai>=1.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Statistical Forecasting
statsmodels>=0.14.0
scipy>=1.11.0

# Prophet (Meta's time series forecasting)
prophet>=1.1.5

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Web Dashboard
streamlit>=1.30.0

# API Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# CLI
click>=8.1.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Typing
pydantic>=2.5.0
